Had help fixing massive issue when using Hibernate with Jackson databind among other random small issues
added a new html file for employee's also js file. I wil most likelt need one for the mangers

Due to the suggestions by Ben to remove the localDate Time from rembursement class, thus removing from from the tables I had to remove the boiler plate code from reimbursement class and redo also adding back in the concatenation in the toString method.

because of removing this I had remove and refactor my html and js because of the missing columns also tweak the class names and moved columns

Informed me that the for loop in th js is named what ever I want  because it is just looping the data so I had to change the name through the cells

due to suggestions by Ben to fix my issue in reimbursement controller and user controller removed the list in the method becuase it returns it in the response error on my end of understanding

I was having same error as kenneth with the issue with hibernate using the object mapper. this was a massive issue presenting me from progressing ... I used the @jsonIdentityInfo  to my class set the property to the id variable name I had to manually import the object Generator....SUCESSS :)
I used these sources below to come  to a solution
the video
https://www.youtube.com/watch?v=j3X88wmtBZQ&t=176s&ab_channel=KindsonTheTechPro
the article based on the video
https://www.kindsonthegenius.com/hibernate-relationship-tutorial-onetomany-and-manytoone-part-8-to/

https://www.baeldung.com/jackson-advanced-annotations

https://stackoverflow.com/questions/3325387/infinite-recursion-with-jackson-json-and-hibernate-jpa-issue

Thank you Ben for helping fix issues with  showing my data in  postman... I somehow some way removed my getters in user class and needed to restart spring tools.

Next I will fix my add reimbursement functionality in my controller, then fix my login, then try my view all users to see if it works and finish up my front end let the struggle  continue :(