99% on the backend now must tighten up the front end
Had a lot of help from mark to make corrections Thank you Mark :)
Also some from Mike as well he always catches the small things

finally added code for the DTO reimbursement thought I wouldnt need it apparently it cant be avoided

added a get status method to reimbursement status

to the sugguestion of Mark I needed more methods in the reimbursement type DAO

needed addition methods in user dao finish get all user method  also added it to the interface and the service

completed with login DTO, srervice I think the controller is okay... havent tested it yet

but I did test the user list in the lanuch it works... some quirkiness to the return but its okay.

I think the backend is okay.... I hope

I need to complete the front end, html, and javascript... fingers crossed I will ask for help as needed but trying to figure it out on my own at first.

I think I am stil missing one method to have the manger change the status of the  reimbursement I started the method in the dao just need to double check it in the controller