Fixed Cors issue I guess
Appaerenly I have to run as a java application then run as a server in order for my tomcat to reconize that I have a cors filter class.

I tried to tidy up the front end

was able to implement view all users

started some of the JS but not complete I have alot to learn to implement what is needed.

nothing is 100 percent because of the issue with the hibernate mapping  so I can only one reimbursement relation is showing and only the first two users relation shows for all user sorry this is the best I can do at this point